requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.0.6
django-admin-sortable==2.2.3
django-cors-headers==3.3.0
django-heroku==0.3.1
django-markdownx==3.0.1
django-storages==1.9.1
dropbox==10.1.2
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.9
importlib-metadata==1.6.0
Markdown==3.2.2
pathspec==0.8.0
Pillow==7.1.2
pkg-resources==0.0.0
promise==2.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dotenv==0.13.0
pytz==2020.1
regex==2020.5.14
requests==2.23.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
whitenoise==5.1.0
zipp==3.1.0
